Refactor table.rs into multiple files
